+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - Added bus scan function to examples/cli.rs
 - Added examples/statusloop.rs
 - Change Rs1541Error to Error
+- new modules for various objects
+- Add [`CbmDriveUnit::try_from_bus`]
 
 ### Changed
 - Moved away from opencbm bindings to A Rust native xum1541 implementation (xum1541)